As-is commit - not working perfectly
Was trying to evolve microscope to the charity deal finder app (Deals
for Good). Ran into challenges with search and reactivity. Became too
complicated with too many things going on in the project.  I'm going to
practice the techniques with simpler codebases and then come back to
this. Saving because I might want to grab pieces from this (like some
of the UI or the data model).